// src/services/axiosApiAppClient/models/OrganizationRole.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum OrganizationRole {
  UNPRIVILEGED = 'unprivileged',
  ADMINISTRATOR = 'administrator',
  GC_CHAIR = 'gcChair',
  GC_TREASURER = 'gcTreasurer',
  GC_SECRETARY = 'gcSecretary',
  GC_PUBLISHING_DIRECTOR = 'gcPublishingDirector',
  DIVISION_CHAIR = 'divisionChair',
  DIVISION_TREASURER = 'divisionTreasurer',
  DIVISION_SECRETARY = 'divisionSecretary',
  DIVISION_LIAISON = 'divisionLiaison',
  DIVISION_PUBLISHING_DIRECTOR = 'divisionPublishingDirector',
  DIVISION_SOP_COORDINATOR = 'divisionSopCoordinator',
  UNION_CHAIR = 'unionChair',
  UNION_TREASURER = 'unionTreasurer',
  UNION_SECRETARY = 'unionSecretary',
  UNION_PUBLISHING_DIRECTOR = 'unionPublishingDirector',
  UNION_SOP_COORDINATOR = 'unionSopCoordinator',
  CONFERENCE_SOP_COORDINATOR = 'conferenceSopCoordinator',
}